agents/Simple Langgraph Agent/requirements.txt
langchain-google-genai
langchain-core
langgraph
tavily-python
python-dotenv